+++ NotesApp/AllNotesScreen/ViewControllers/AllNotesViewController.swift
@@ -11,14 +11,16 @@ class AllNotesViewController: UIViewController {
     
     var palette: ColorSet
     var contentView: AllNotesView
+    var collectionDataSource: AllNotesDataSource
     
-    init(palette: ColorSet) {
+    init(palette: ColorSet, collectionDataSource: AllNotesDataSource) {
         self.contentView = AllNotesView(palette: palette)
         self.palette = palette
+        self.collectionDataSource = collectionDataSource
         
         super.init(nibName: nil, bundle: nil)
         
-        self.contentView.collectionView.dataSource = self
+        self.contentView.collectionView.dataSource = collectionDataSource
         self.contentView.delegate = self
     }
     
@@ -54,51 +56,51 @@ class AllNotesViewController: UIViewController {
     }
 }
 
-extension AllNotesViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        if section == 0 {
-            return 6
-        } else {
-            return 10
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        if indexPath.section == 0 {
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: TagCollectionViewCell.identifier, for: indexPath) as? TagCollectionViewCell else {
-                fatalError()
-            }
-            cell.setup(with: .init(colorSet: .classic, tag: "#teste"))
-            return cell
-        } else {
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: NoteSmallCellCollectionViewCell.identifier, for: indexPath) as? NoteSmallCellCollectionViewCell else {
-                fatalError()
-            }
-            cell.setup(with: .init(colorSet: .classic, title: "Oi", content: "ashcasodha"))
-            return cell
-        }
-    }
-    
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 2
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        if indexPath.section == 0 {
-            guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: TagHeader.identifier, for: indexPath) as? TagHeader else {
-                fatalError()
-            }
-            header.setup(with: "Tags")
-            return header
-        } else {
-            guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: NoteHeader.identifier, for: indexPath) as? NoteHeader else {
-                fatalError()
-            }
-            header.setup(with: "All Notes")
-            return header
-        }
-    }
-}
+//extension AllNotesViewController: UICollectionViewDataSource {
+//    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+//        if section == 0 {
+//            return 6
+//        } else {
+//            return 10
+//        }
+//    }
+//
+//    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+//        if indexPath.section == 0 {
+//            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: TagCollectionViewCell.identifier, for: indexPath) as? TagCollectionViewCell else {
+//                fatalError()
+//            }
+//            cell.setup(with: .init(colorSet: .classic, tag: "#teste"))
+//            return cell
+//        } else {
+//            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: NoteSmallCellCollectionViewCell.identifier, for: indexPath) as? NoteSmallCellCollectionViewCell else {
+//                fatalError()
+//            }
+//            cell.setup(with: .init(colorSet: .classic, title: "Oi", content: "ashcasodha"))
+//            return cell
+//        }
+//    }
+//
+//    func numberOfSections(in collectionView: UICollectionView) -> Int {
+//        return 2
+//    }
+//
+//    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+//        if indexPath.section == 0 {
+//            guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: TagHeader.identifier, for: indexPath) as? TagHeader else {
+//                fatalError()
+//            }
+//            header.setup(with: "Tags")
+//            return header
+//        } else {
+//            guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: NoteHeader.identifier, for: indexPath) as? NoteHeader else {
+//                fatalError()
+//            }
+//            header.setup(with: "All Notes")
+//            return header
+//        }
+//    }
+//}
 
 extension AllNotesViewController: NoteSmallCellCollectionViewCellDelegate {
     func didTapDelete(for noteViewModel: SmallNoteCellViewModel) {
